ttools: refactor Ganger/GangerFactory API
The Ganger now knows how many zones it has, so no methods on it need
to pass that value in, and the GangerFactory can accept a ConfigMap
with advertised ConfigKeys to control the Gangers that it produces.
A new class GangContext is required that provides plot-time
information to the GangerFactory about the gang that it will need
to produce.

The GangerFactory is now available from the PlotType, and topcat's
GuiPlotType does not need to make additional arrangements to provide it.