topcat: TopcatLayer is now per-plot not per-zone
The TopcatLayer now contains an array of PlotLayers, one for each
zone in the gang, rather than a single PlotLayer.
TopcatLayer is now therefore a per-plot object (going with a Gang)
rather than a per-zone object.

TopcatLayer instances are still supplied to the StackPlotWindow by
LayerControls, but a Ganger must now be presented in order to get them.
The TopcatLayer array is transmitted by StackPlotWindow to PlotPanel
using a new Supplier argument in the PlotPanel constructor rather
than in the ZoneDefs, since a TopcatLayer is no longer per-zone.

In all existing cases, a TopcatLayer contains a single PlotLayer
(only one element of the nzone-element PlotLayer array is non-null),
as implemented in the new classes SingleZoneLayerControl and
SingleZoneLayerManager.

However, future layer controls may be written which exploit the
new capabilities and have multiple PlotLayers per TopcatLayer.

Behaviour is not supposed to change at this commit, but there has
been quite some reorganisation, so regression issues are possible.