ttools: allow global or per-zone trim and shading.
Ganger has new methods isTrimmingGlobal and isShadingGlobal.

According to these, the Trimming and ShadeAxisKit arrays supplied
when making a plot are no longer constrained to be the same length
as the zone count.  Either can be either per-zone (nz-element array)
or global (1-element array).  That means that the legend and/or
shade axis can be applied to either the whole multi-zone plot,
or one per zone.

In topcat, AxisController is replaced by AxesController.
The name change isn't important, but the new class has a set of
ZoneControllers, so that the AxesController is no longer required
to manage a single Zone, in principle it can manage multiple zones.
At this commit the AxesController implementations are based on
wrappers around the legacy AxisController classes, but it might
be tidier in future to re-implement them directly.

In topcat the communications between StackPlotWindow and PlotPanel
are reworked as well; there is now a single PlotContent object
used to encapsulate all the necessary information from the GUI
and passed to PlotPanel.