ttools: expose some functionality for external use
Make some private methods public, and split out the
ConfigParameterFactory interface out as a separate class.

Also make it a bit clearer what kind of multi-zone plots the
methods in this class can cope with; it can't do global trimming
or shade axes (though subclasses could be written that can).