ttools: apply gang profile/aspect preparation correctly
Ensure that the adjustProfiles and adjustAspects methods of
Ganger are called when they should be when preparing plots in
STILTS and utility classes.

This was not being done correctly before; one consequence was that
the horizontal (time) axis on stacked (time) plots was labelled
for each plot rather than only once on the bottom.