topcat: localise coord fill-in logic
Move a hack that was required to deal with which coordinates need
to be filled in during CoordPanel autopopulation from the CoordPanel
itself to the XYArrayCoordPanel which is where it's actually
required.  This is more general since subclasses can do the
appropriate thing and it doesn't require special-casing applied
at the top level.

The intention is that there is no change of behaviour for
any of the existing classes.