topcat: belatedly acknowledge Europlanet support
Add a record of the support from the Europlanet 2024 RI project
which provided one month of funding in 2021.
The text in the Acknowledgement section of SUN/253 is updated,
and a truncated version of the Europlanet Society logo (edited from
https://www.europlanet-society.org/wp-content/themes/epn2020/img/logo.svg)
is added to the logo panel in the Help|About TOPCAT window.

This should have been done in 2021 but I forgot; reminded by
Stephan Erard on 21 Sep 2023 who noted:

   The standard sentence is The Europlanet-2024 Research Infrastructure
   project has received funding from the European Union's Horizon 2020
   research and innovation programme under grant agreement No 871149.
   + EU flag
   —  I guess just adding it to your list would be OK