topcat: improve CheckBoxList
Tigten up the generics in CheckBoxList, the rendering is now
genericised and easier to customise.

I did this in preparation for an additional use of CheckBoxList
that requires more flexibility.  I may not end up using it there,
but this leaves the code in an improved state in any case.