ttools: rename plot2matrix -> plot2corner
Audience research suggests that "corner plot" is the most widely
known term to describe this kind of plot.  The python package that
people use is called corner.py.  I don't like it that much,
partly because depending on configuration this one might not end
up restricted to the corner, but user recognition/comprehensibility
is probably more important.