topcat: BasicCheckBoxList.setCheck notifies listeners
The setCheck method of BasicCheckBoxList now notifies ListDataListeners
that a change of list entry has taken place.
It's not clear from the abstract superclass documentation that it is
supposed to do this, but if it doesn't then programmatic invocations
of setChecked are generally not noticed or reflected in the GUI.

This is misdesign in CheckBoxList (which is a JList but maintains
ListDataListeners, itself not really right since that ought to be
the preserve of the ListModel).
I should really clean the whole situation up by rewriting some parts
of CheckBoxList or at least its documentation, and ensuring that
its existing subclasses do the right thing, but for now this fix
is better than nothing.