vo: use ADQL-2.0 not 2.1 as default in TapCapabilityPanel
The TAP window was using ADQL-2.1 as the default VersionedLanguage
if the available Capabilities declared no ADQL versions.
That's bad, since a server that isn't declaring capabilities
properly is probably not very advanced, i.e. likely ADQL 2.0
rather than 2.1 compliant; in practice this meant that when talking
to TAP services that did not declare supported ADQL versions,
TOPCAT would probably send LANG=ADQL-2.1 which the service would
probably reject as an unrecognised language.

Fix it so that the default is ADQL 2.0.