topcat: work round JComboBox performance misfeature
In Oracle/OpenJDK8 at least, the JComboBox fires an ActionEvent
when the ComboBoxModel changes but the selected item is null
and does not change.  The expectation is that an ActionEvent is
only fired when the selected item changes (which is true for
non-null values).

This causes a major performance issue in some cases, since there
may be a very large number of ColumnDataComboBoxes all listening
to the same ColumnDataComboBoxModel (which is based on a
TableColumnModel).  If the TableColumnModel changes, then
every ColumnDataComboBox representing it which has a null
selection will receive an action.  If you do something like delete
all columns or remove a low-numbered column in a TableColumnModel,
it will generate a number of events comparable with the number
of columns.  So if there are unselected ColumnDataComboBoxModels
you can get them responding to a huge number of useless ActionEvents.
Each of these will trigger a replot, which is intentionally a
rather cheap action but it's not free, so this can cause genuine
responsiveness issues.

Example: opening a Matrix Plot (which in its current implementation
maintains a load of invisible ColumnDataComboBoxes in its
MatrixPositionCoordPanel) on a wide table and then hitting the
"Hide All" button in the Columns window.  For 80 columns this
takes (e.g.) 10 seconds on the Event Dispatch Thead at 100% CPU.

This commit works round this by basing ColumnDataComboBox on a
modified implementation (subclass) of JComboBox, FixedJComboBox.
It does an additional test (for null equivalance) when the
selection changes before passing it on to the superclass
to fire an ActionEvent.  This seems to fix the problem.
Note it does it by overriding a method on JComboBox that warns
"do not override" but I don't *anticipate* problems.

Note: there are other problems related to listeners and events
in the TOPCAT GUI; listeners are not released when they should be
so there are too many of them, especially hanging on to the
TableColumnModel.  At present this does not cause known problems
in practice, but it would be good to do something about it in
future.