topcat: update multiwindow title-page screenshot
A Matrix plot example is added on the multishot.
In the full-size version, the background is white rather than
transparent since otherwise Fop doesn't process it for the PDF
output (there's probably a way round this but I haven't worked
out what).
The mini version got a Gaussian Blur from gimp before shrinking
to 1/4 linear dimension.  This doesn't look as good as last time
(it's fainter) but the blur I used then seems to have disappeared
from gimp.