ttools: fix BinMapper overflow bug
Fix a bug in linear and log BinMapper construction that could
encounter overflow when calculating reference point for bin indices.
The effect was an AssertionError (when -enableassertions) or
something worse like an OutOfMemoryError while trying to plot
a histogram.  This didn't happen often, but could do if the
range of histogrammed points was small (e.g. one point only)
and far from the origin.

I was for no good reason casting the output of Math.floor to (int),
so I stopped doing that.

Some kind of overflow is probably still possible for very extreme
referencePoint/range ratios, but I'll deal with that if somebody
reports it.

This bug was spotted by Benjamin Steltner (MPG).