ttools: formatPosition reports panel in Matrix plots
The PlanePlotSurface.formatPosition can now be configured to include
the axis labels alongside the numeric values for the reported
data position.  This is helpful for the Matrix Plot, since it may not
otherwise be easy to see by looking which coordinates the panel
you are pointing at is representing.

The configuration is done using an additional method of the
PlaneSurfaceFactory.Config interface, set false for contexts
other than the Matrix Plot.