vo: provide pre-emptive authentication in TAP window
An action is now provided for authenticating in the TapTableLoadDialog.
This can be activated via a button in the window, and it may also
be forcibly activated when a new TAP service is contacted,
if authentication is required for use of that service.

The action contacts the /capabilities endpoint in accordance with
the SSO_next proposal to establish authentication requirements and
perhaps to authenticate; TapServiceKit uses the authcheck method
of AuthManager to do this.

This means that the TAP service can effectively be contacted
pre-emptively to ensure that the correct metadata is acquired
for the (maybe) authenticated connection status that will be
used for communicating with it.

This relies on proposed behaviour currently only documented at
https://wiki.ivoa.net/twiki/bin/view/IVOA/SSO_next.
Changes may therefore be required as the standardisation process
advances.