ttools: adjust plot2corner documentation
A couple of the example invocations for the plot2corner task had
the wrong parameters; not broken, but excessively boring and not
matching the description text.  Do some other minor rewording too.

Also add an example graphics output image to the plot2corner task section.
The other plots don't currently have this, but it's really required
for plot2corner since (a) there is no other matrix plot screenshot
elsewhere and (b) it's worth a thousand words given the plot type
is hard to describe.

I should in the future homogenise the plot type sections,
e.g. by adding a figure to them all, or maybe incorporating
screenshots in the examples sections.