ttools: fix votlint for empty TDs at VOTable 1.3+
Up till now, votlint would report an error for empty TD elements
if the FIELD data type was not null-friendly (i.e. if nulls could
not be represented in the BINARY serialization); this applied to
integer and fixed-length array types.

That was correct behaviour for VOTable versions up to and including
version 1.2, but for version 1.3 and later empty TD elements are
permitted for all data types with the meaning of a null value
(see VOTable 1.3 section 5.5).

Following this commit, an error is reported in this situation
only for VOTable versions 1.2 and below.

I am surprised that this incorrect behaviour had gone unnoticed
for so long; VOTable 1.3 is now ten years old.