ttools: fix tcat/tcatn bug
If tcat or tcatn was trying to catenate a table with an indeterminate
row count and additional columns (any of the parameters seqcol,
loccol or uloccol set), then no rows were included.
The problem was to do with creating a constant-column-valued table
using ColumnStarTable but that was not really random-access.
Now fixed.

This bug was reported by Benjamin Steltner (MPG).