task: allow multiline entries in indirection files
When a task parameter is retrieved from an indirection file
(specified with param=@filename), interpret trailing backslashes
in lines of the file to ignore the following newline character.
In this way entries can be split over multiple lines.

The parsing is not very sophisticated, and no other interpretation
is done at this stage, so for instance it is now impossible to
specify an entry that actually contains a trailing backslash.
But this usage is sufficiently obvious that I think it's reasonable
to implement, it's expected behaviour, and I don't expect it to
trip anybody up.