topcat: provide Function layer control from MatrixPlot
The function layer is already available for the MatrixPlotType
from stilts, but wasn't from topcat because the layer control GUI
did not provide a multizone-friendly interface.
Rewrite it so it does, which was pretty straightforward.

There is however a loose end.  As it stands, the function layer
appears in the same form on all gang zones.
In practice, it's unlikely that you will want the same line to
be displayed on both off-diagonal (scatter-plot-like) and on-diagonal
(histogram-like) panels.  So really there should be some user
interface in both stilts and topcat to select in the case of the
matrix plot where you want the functions to appear (on- or off-diagonal).
More generally you could imagine wanting different functions in all
different panels... but I don't want to get into that.

For now leave it in its current state until somebody complains.