ttools: Coverage functions can use ASCII MOC literals
The first (string) argument of the inMoc and nearMoc functions
can now be an ASCII MOC string as an alternative to the location
of a FITS MOC file or the identifier of a VizieR table.

I added some new tests for the older MOC functionality as well as
for the new parts.
The MOC file added to the testcases directory was downloaded from
http://alasky.u-strasbg.fr/footprints/tables/vizier/II_321_iphas2/MOC