ttools: add -xtype flag to metadata filters
For some reason it has not been possible to edit the xtype metadata
attribute of table columns and parameters using stilts up till now.
Fix this by adding a -xtype flag to filters addcol, colmeta,
replacecol and setparam.  This was an easy upgrade to make:
usage and implementation follow exactly the form of the existing
-utype flag.