ttools: ISO-8601 date conversions accept trailing Z
The Times.isoTo*() functions accepted a trailing Z for date-time
input strings, but not for date-only ones.  Now they do both.
Do the same thing for table.TimeMapper.

I don't have access to the text of ISO-8601 itself, but related
documents (e.g. RFC 3339) suggest that a timezone designator
such as Z the is only permitted as part of the Time part and
not with a Date only.
However, in at least some cases a Z is simply slapped on after
whatever date-time/time/date string is present, and the
functions here are for comprehending existing text not for
validation, so it makes sense to permit this regardless of
the intention of the standard.