fits: fix bug reading multiple table HDUs
When reading multiple tables from a FITS stream, it is necessary
to determine when the stream is ended.  The existing code didn't
do this correctly (at least in some cases) so could lead to an
EOFException("FITS ended before header") when it shouldn't have.
For instance this generated a non-fatal error message when using
stilts tmulti or topcat load on a fits-basic, though not fits-plus,
file.

Rework the code a bit to avoid the pre-emptive EOF detection code
which was hairy and self-confessedly relied on flawed parts of the
JRE I/O classes.