ttools: report geometry parameter in task autodocs
Plot2Tasks optionally have parameters for determining DataGeom,
but up till now such parameters were not reported by the Task,
hence not included in the auto documentation.

That meant that plot2sky's datasysN parameter was not documented.
It also meant that there is an undocumented "geom" parameter
I had long forgotten about that could be used to change the
DataGeom per-layer for other plot types, so that for instance:

   stilts plot2cube in=:loop:1000 \
                    layer1=mark geom1=polar lon1='i%360' lat1='45' r1=i

produces a plot.

This commit fixes it up so that this is now documented.

Only plot2cube and plot2sky are affected, since the other PlotTypes
have a single fixed DataGeom, so no DataGeom choosing parameter
is useful.