ttools: clarify NULL_ evaluation in docs
Add a note in the expression language sections of SUN/265 and SUN/253
that NULL_ tests should come first in short-circuit expressions.

The reason is that attempting to dereference a null primitive causes
immediate termination of expression evaluation, while the NULL_
test returns a value without such termination.  I haven't included
that explanation in the text because it requires understanding of
certain concepts the reader might not have, and would complicate
the text quite a bit.