votable: support COOSYS/@refposition for VOTable 1.5
The refposition attribute of the COOSYS element, introduced in
VOTable 1.5, is now incorporated into STIL metadata on VOTable read,
which means it can be examined in TOPCAT and STILTS under the
Aux metadata heading "CoosysRefposition".
It is also written out to VOTable when present if the VOTable output
version is >= 1.5.