fits: improve error reporting for corrupted FITS files
Following a user report about a cryptic error message that took
me a while to track down, I have made checks and reports more
explicit for the case when it can be determined that FITS headers
imply HDU lengths that overrun the end of their file.
It should now be more obvious at an early stage that a corrupted
or truncated FITS file has been encountered.