vo: use RESPONSEFORMAT not FORMAT for TAP 1.1
TAP 1.1 sec 2.7.3 recommends use of the DALI-compliant
RESPONSEFORMAT parameter to specify preferred response format,
but TAP 1.0 uses the name FORMAT to mean the same thing.
TAP 1.1 is required to accept FORMAT as an alias.

Adjust the code so that services known to be TAP 1.1 get
RESPONSEFORMAT, and only services that may be TAP 1.0 get
the older FORMAT.