votable: use auth for href-indirected STREAM input
Fix the data reads for STREAM data that come from external URLs
(rather than being inlined and base64-encoded) so that they
use the AUTH library.  This means that they can work even from an
an auth-protected URL.  It also provides HTTP-level GZIP encoding.

The GZIP encoding is hardwired in at this commit.  Possibly it
should be configurable whether it's used or not, but I don't
really know why you'd want to turn it off.