vo: allow choice of TAP metadata tree ordering
There are now radio buttons near the TAP schema/table metadata
tree display that let you choose whether the metadata is to be
ordered alphabetically (as before) or by service-provided
table/schema index.

This is potentially helpful, except that at time of writing
for most services the table/schema index is not filled in
at all or with much thought.