vo: improve block scrolling in TAP window MetaPanel
The getScrollableBlockIncrement method for vo.MetaPanel was
implemented the same as getScrollableUnitIncrement, which meant
that clicking on the scroll bar trough scrolled only a line at
a time.  Fix it so that it does about a screen at a time.
This isn't perfect, it would be nice if the scroll guaranteed
to put a line break at the top of the page each time, but that
would be rather complicated and it's not a big deal.