ttools: add Vincenty formula for angular separation
Add a non-public method vincentySeparationFormula to CoordsRadians
alongside the existing haversineSeparationFormula.  These do the
same thing, but Vincenty is stable at angles near Pi, where
Haversine blows up.  Vincenty is about 50% slower.
Vincenty is not strictly symmetric in its arguments, so one test
has to be changed to accommodate differences of order 1e-15 in
a comparison in the case that Vincenty is used.

A static flag USE_VINCENTY determines which of the two formulae
is used by the user-visible skyDistanceRadians function.

At this commit USE_VINCENTY is false, so there is no change in
behaviour.  But maybe I'll turn it on one day.