vo: refactor TableSetPanel configuration
Modify the TableSetPanel and TapQueryPanel constructors so that
they take a TapTableLoadDialog rather than a UrlHandler.
This is because they are going to need more configuration,
and it's convenient for them to take multiple items from the
owning TapTableLoadDialog rather than to increase the number
of bespoke constructor arguments.  No behaviour change.