vo: improve TAP table metadata tree scrolling
Fix various issues with the way that the metadata display tree
behaves when you try to scroll it to a given path.
If possible make both the path and its parent both be visible,
and deal with the, probably rare, case that the selected path is
currently masked out of the tree by a search filter
(by resetting the search filter).