vo: propagate TAP finder selection to metadata display
If the user selects a table in the service finder tree,
take some steps to highlight this selection in the metadata
display tree in the query panel.

Specifically: the first time the service metadata is displayed,
and if the user subsequently double-clicks on the table in the
service finder tree, set the metadata tree selection to use
the table selected from the service finder tree (if any).

The keyword search filter is also no longer reset when a new
set of service metadata is installed.  That was a bit annoying,
especially when waiting for the initial metadata population,
and I don't really know why it did it.