ttools: fix TZ bug in TimeConfigKey formatting
Incorrect handling of the notoriously tricksy Date/Calendar/DateFormat
classes meant that the formatting of time coordinate values into
ISO-8601 strings was time-zone dependent (and it showed a 1-hour offset
from UTC in the UK in the winter, which I still don't understand).
This meant amongst other things that the time axis for plots in
stilts commands generated from topcat plots could be some number
of hours offset from where they should be.  Fix it by making sure
that everything is done in UTC.