topcat: fix layer config for spectrogram
Provide an implementation for DisjointAxesController.configureForLayers.
Through laziness (I didn't need it yet and wasn't sure how to do it)
this implementation was left as a no-op when the AxesControllers
were reorganised to accommodate the matrix plot a few months back,
but it's needed now for adjusting spectrogram axis labels,
so do it right.