ttools: spectrogram plot now tries to use spectral axis
The SpectrogramPlotter now tries to find some metadata that looks
like it defines the channel grid (mapping of spectral elements onto
the vertical axis) and uses it if available.  Previously it just
used array index for this mapping.  This search can be turned on
or off via a config key, but is not otherwise configurable
(partly since there is no way to reference column metadata items
in the expression language), and a bit ad hoc (it grabs any array
in the column/table metadata that seems to have the right dimensions),
but it will probably do the right thing more often than not.