topcat: fix stupid but apparently harmless code
Assembling a list of ZoneControllers in StackPlotWindow
was generating an array of nulls, with some dead code that
looked like it did something but didn't.  However, as far as
I can tell, having an array of nulls here wasn't causing anything
not to work.

The code now at least looks like it's doing something sensible,
though whether the effect on behaviour is positive, negative or
non-existent I'm not yet sure.