ttools: use non-default text style in plot example
The layer-polygon example plot now has LaTeX-style axis labelling
at a non-default size.  This is partly to demonstrate in the
user documentation that this can be done, and partly to provide
a test of this functionality in the stilts UI, since the images
produced for these examples are compared for regression as
part of the release process.