topcat: fix caption config bug in Time Plot
In the (multi-zone) TimePlotWindow, adjusting the caption style
configuration in the AxesController Font tab was correctly updating
the style of the plot axes, but not affecting the caption style
of the legend or aux ramp labels (which kept the default style).

I have fixed this up by acquiring the relevant configuration items
directly from the axes controller and passing them to the legend
and ShadeAxis creation routines explicitly, rather than trying
to get the LegendControl and ShadeAxisControl to manage this
zone-specific configuration themselves.  This is more robust as
currently implemented, though maybe less elegant.

This zone-specific part of the code (largely fallout from the
recent introduction of the Matrix plot) is a bit of a mess;
I don't really know if the MultiConfigger does anything or where
all the zone configuration items come from.  The code could benefit
from a careful clearout.  But for now I think everything is working,
so leave it be.