ttools: fix StiltsPlot matrix plot profile key omission
The STILTS control in the topcat plot window was failing to pick up
some of the config keys for the MatrixPlot.  The effect was that
if you set things like font and grid characteristics in the topcat GUI,
although they showed up in the display, the stilts command they
generated did not include those settings.
That happened because StiltsPlot handles the MatrixPlot profile keys
specially, but was only taking care of the X/Y-specific ones,
and the non-geometric profile keys had fallen through the cracks.
Reinstate them.