vo: present UDFs in a separate tab of TAP window
The User-Defined Functions declared in TAPRegExt are now presented
in a JTree in a dedicated tab in the TapQueryPanel rather than
being squashed in at the end of the Service tab.
Services are starting to offer long lists of these with involved
documentation now so it makes sense to give them more space.