topcat: allow zone gap configuration for Time plot
The Frame control now offers a new "Spacing" tab for the Time Plot
only.  This allows configuration of the inter-zone spacing for
stacked plots.  Implementing this requires that the Frame config
is distributed to certain places that it wasn't before.

This arrangement isn't ideal; this item would fit more naturally
into the Axes control, but that is currently managed per-zone
for the time plot, and the spacing configuration is per-window
not per-zone.