topcat: Time Axis label can now be configured in Time Plot
The labelling option was deliberately absent for the time axis on
the grounds, presumably, that the Time axis is the Time axis.
But it's harmless and sometimes desirable to be able to
put a custom label on the horizontal axis of a time plot,
including if you are abusing the time plot to plot non-temporal
quantities.