vo: choose Service table/schema ordering as default
Use service-specified rather than alphabetical ordering as the
default setting for the TAP metadata tree display.

I originally decided that alphabetic was better on the grounds that
when you get metadata from the /tables endpoint (VOSI) you get an
ordering that is defined, but you can't tell whether it was intended
by the service (unlike for TAP_SCHEMA when it's obvious whether they
have filled in the table_index/schema_index fields), so you may
end up presenting the tables/schemas in an essentially random order.
But following some investigation it seems that most (though not all)
services list tables in the VOSI TableSet which is either alphabetical
anyway, or appears to have some thought behind it.

So I think on balance using service-defined ordering as a default
is preferable.