ttools: improve topcat and stilts startup scripts
Make the topcat and stilts startup scripts more robust to embedded
spaces in the pathname of the script directory, and revise the
behaviour on MacOS in view of the recently changed structure
of the DMG bundle in which the MacOS application is generally
distributed.  They now have a better idea of where nearby resources
are located, and use the bundled JRE to run.

Also update the advice about invoking stilts from the command line
in SUN/256.