tjoin: incorporate errors-in-quadrature matcher
Move the functionality of ErrorQuadratureCartesianMatchEngine
into ErrorCartesianMatchEngine, which now has an additional
construction-time parameter to choose whether behaviour should
be with or without quadrature.  This is more compact and leads
to less duplicated code.

Make the with-quadrature options available from the topcat and
stilts UIs, and update the documentation accordingly.

This option was requested by Alexey Mints, who contributed a
version of the code.  Making this addition is a bit unsatisfactory,
since the with-quadrature option is now available for
Cartesian-with-errors matching, but equivalent options are not
provided for sky-with-errors or Cartesian/sky elliptical matcher types.
I'm reluctant to do the same thing for all of these because
the list of possible options in the UIs, especially TOPCAT's GUI,
would get more bloated and become harder to use.