ttools: retire obsolete taplint DALI-examples logic
Because of standards confusion in early DALI versions,
it was impossible to write a DALI-examples document that was
standards-compliant and that TOPCAT v4.3-4.4 would recognise.
Taplint therefore had complicated logic and messages advising
how/whether to work round this.

Since this problem is fixed since TOPCAT v4.5 (Sep 2017) these
warnings are now considered obsolete, so tidy things up a bit,
at least in terms of making the advice to users less confusing
(don't bother attempting to cater to old TOPCAT versions in
this respect).