vo: set ADQL JTree row height to zero
Make an explicit call to setRowHeight(0) for the JTree displaying
ADQL language features in the TAP window.
A row height of (<=) zero is required to make variable-height cells
in a JTree work, and some of the cells here may have multiple lines
of text.

On my linux development platform the default configuration worked
fine (rowHeight was zero by default) but it seems this wasn't the
case on MacOS (OpenJDK and Oracle java 8 on ARM at least) so only
the first line of such cells was being displayed.  With this commit
it now works for both platforms.