vo: fix HTML issues in ADQL Hints panel
The content of the Hints tab in the TAP window was failing to load
for Java 11 (though not Java 8) because reading the HTML document
constructed by XSLT at build time was failing with a
javax.swing.text.ChangedCharSetException.

I didn't get to the bottom of what that exception is for or why it's
thrown under those circumstances, but modifying the XSLT invocation
to write X(HT)ML rather than HTML seems to sort it out.