ttools: HandlerDoc now reports bean default values
Add code to HandlerDoc that tries to find out and report the
default values for the configuration options that it documents.
This requires bean-like accessor methods (not currently present
in all cases) alongside the existing bean-like mutator methods
with which HandlerDoc reflectively interacts.

The effect is that users can see from the auto-generated documentation
what the default values are for the various I/O handler configuration
options; this information was not previously visible from the
documentation.