parquet: update to parquet-mr 1.13.1
The main reason for this update was that the snappy compression
implementation in parquet-1.11.1 (provided by snappy-java-1.1.7.3.jar)
did not contain Mac/aarch64 native libraries, so that parquet files
with snappy compression (generally the default) were not readable
on ARM-architecture Macs.  The tidiest approach to that would be
to build or otherwise acquire the requisite libsnappyjava.dylib
binary and add it to the existing snappy jar file; but I don't
know how to do that.  So I upgraded the parquet library classes
to a parquet-mr version in which the aarch64 native files are
present by default.  Possible parquet versions were 1.12.3 and
1.13.1; I chose 1.13.1 since it's more recent but not massively
more bloated.  It's possible that this version update might have
other benefits (or problems) too.
Another approach would have been upgrading just snappy to 1.1.8.3
and keeping the rest of parquet-mr at 1.11.1, but that seemed
a bit risky.

The update was done following the procedure for the initial import
of parquet-mr 1.11.1 when the PARQUET package was introduced.
Specifically:

   - clone https://github.com/apache/parquet-mr
   - checkout tag apache-parquet-1.13.1
   - cd parquet-cli
   - mvn clean install -DskipTests   // probably optional
   - mvn javadoc:javadoc             // optional
   - mvn dependency:copy-dependencies

The required jar files can then be found in parquet-cli/target and
parquet-cli/target/dependency.

Not all of the declared parquet-cli dependencies are needed by the
STIL parquet I/O handlers, so to avoid more bloat than necessary
some trial and error was used to determine the minimal set of
library jars required for STIL I/O using common compression formats,
which are those included in this commit.  It is possible that I
didn't do this trial and error thoroughly enough, i.e. that some
code paths will require jar files I haven't included in this commit,
so later revisions might be necessary.  Even after this culling,
parquet/src/lib still contains 21Mb of library jars, which is more
than the whole of STILTS and all of its other dependencies.
This is annoying.

Some subsequent tidying up of jar files may be required to ensure
that the monolithic jar file build (topcat build-standalone target)
works properly.