parquet: add parquet LZ4_RAW support
LZ4_RAW (a.k.a. LZW) support comes from the pure-java aircompressor
library (https://github.com/airlift/aircompressor).

This library looks really useful, since it seems to contain a
pure-java snappy implementation, including a Hadoop CompressionCodec;
this ought to be able to replace the JNI-based org.xerial.snappy
implementation currently pulled in from snappy-java-1.1.8.3
(it claims to be faster as well).  Looks like it has zstd and lzo too.

On the other hand... the README.md says
"This library requires a Java 1.8+ virtual machine containing the
sun.misc.Unsafe interface running on a little endian platform."
which sounds a bit scary.

But I so far have not figured out how to configure the implementation
classes for given compression formats in parquet-mr :-(.