ttools: improve JELRowReader typed evaluation
The JELRowReader evaluateDouble method will now return a double
primitive for expressions with any numeric type, not just double;
the documentation suggested this behaviour but it's not what it
was doing up till now (it returned zero with a log message).
Update the evaluateBoolean method similarly, so that attempts
to evaluate a non-boolean expression behave better.