tjoin: add skyerror-in-quadrature matcher
Alongside the errors-in-quadrature matcher, implement the same thing
for skyerror.

At the same time, withdraw the errors-in-quadrature option from
the TOPCAT GUI, just leave it in the STILTS UI.
Putting too many of these things in clutters up the TOPCAT GUI too
much, and it's a subtle and somewhat niche distinction.
Having both variants available is less obtrusive in STILTS,
and Alexey Mints who requested it said that would be a reasonable
compromise.  If people complain that this functionality ought to be
available in TOPCAT as well, deal with it then.