ttools: allow arbitrary quantiles in Combiner parameter
Replace the use of ChoiceParameter<Combiner> with the new
specialised class CombinerParameter.  This does the same thing
but with an additional feature: arbitrary quantiles can now be
specified using the same syntax as for the stats filter,
so Q.01 represents the first percentile etc.
This applies also to the combiner choices in the
Grid and Sky density map column specifications.