ttools: note taplint complies with REC-EPNTAP-2.0
There's no change in behaviour here, but EPNTAP 2.0 has become REC
(some while ago, 20220822) since the source code was last changed,
and I have verified that there are no relevant changes in the
text since the last version that I checked compliance with (2022-07-04),
so adjust the comment in the EpnTapStage javadocs to declare
compliance with REC-EPNTAP-2.0.

Probably such version number compliance should be included per-standard
in the taplint user documentation, but at present it is not.