ttools: redo ASCII MOC identification
Change the way that a string is tested to determine whether it
looks like an ASCII MOC string in the Coverage class.
This was previously done using a regular expression, which was
(as far as I know) correct, but prone to cause StackOverflowErrors
during regex matching when confronted with long test strings.
It might be possible to avoid this by using possessive quantifiers
in the regex, but my (somewhat desperate) experiments in this
direction didn't work.

So do this by hand.  The relevant method is moved to the
AsciiMocCoverage class since I want it available for use elsewhere,
and it's not appropriate as a user-visible function in the
expression language.

The implementation is currently rather inefficient and could
be improved, but it's probably not going to be called all that
often.