ttools: EPN-TAP coverage xtype is "moc" not "MOC"
Earlier versions of EPNTAP-2.0 required xtype="MOC" for
the "shape" field, but in the final REC-EPNTAP-2.0 this field
has become "coverage" with xtype="moc", which also follows the
current-at-time-of-writing presecription in WD-DALI-1.2-20230712.
I must have missed the xtype case change in this case during
document evolution.  Fix it here.

Most epn_core tables out there actually do the right thing
(xtype="moc" for the coverage field), though a few have xtype="MOC",
so happily people have been (mostly) coding to the standard and not
to the validator, which was wrong in this case.