ttools: fix potentially fatal votlint bug
There was an input-byte-counting bug that meant if votlint was run
on a BINARY- or BINARY2-serialized VOTable with a column having
arraysize="*" and xtype="polygon" the read would lose track of
its position in the binary stream and most likely fail with an
Error report saying "Unrecoverable stream error".

Fix it.